+++ src/Spice/Core/DatePicker.cs
@@ -7,7 +7,6 @@ namespace Spice;
 /// </summary>
 public partial class DatePicker : View
 {
-
 	/// <summary>
 	/// The selected date
 	/// </summary>
@@ -37,10 +36,10 @@ public partial class DatePicker : View
 	{
 		base.ApplyTheme(theme);
 		_isApplyingTheme = true;
-		if (CanApplyTheme(ThemeProperty.TextColor))
+		if (CanApplyTheme((int)ThemeProperty.TextColor))
 			TextColor = theme.TextColor;
 		_isApplyingTheme = false;
 	}
 
-	partial void OnTextColorChanging(Color? value) => TrackExplicit(ThemeProperty.TextColor, value);
+	partial void OnTextColorChanging(Color? value) => TrackExplicit((int)ThemeProperty.TextColor, value);
 }
