+++ src/Spice/Core/Label.cs
@@ -7,7 +7,6 @@ namespace Spice;
 /// </summary>
 public partial class Label : View
 {
-
 	/// <summary>
 	/// The text to display
 	/// </summary>
@@ -25,10 +24,10 @@ public partial class Label : View
 	{
 		base.ApplyTheme(theme);
 		_isApplyingTheme = true;
-		if (CanApplyTheme(ThemeProperty.TextColor))
+		if (CanApplyTheme((int)ThemeProperty.TextColor))
 			TextColor = theme.TextColor;
 		_isApplyingTheme = false;
 	}
 
-	partial void OnTextColorChanging(Color? value) => TrackExplicit(ThemeProperty.TextColor, value);
+	partial void OnTextColorChanging(Color? value) => TrackExplicit((int)ThemeProperty.TextColor, value);
 }
