+++ src/Spice/Core/Picker.cs
@@ -9,7 +9,6 @@ namespace Spice;
 /// </summary>
 public partial class Picker : View
 {
-
 	/// <summary>
 	/// The collection of items to display in the picker
 	/// </summary>
@@ -54,10 +53,10 @@ public partial class Picker : View
 	{
 		base.ApplyTheme(theme);
 		_isApplyingTheme = true;
-		if (CanApplyTheme(ThemeProperty.TextColor))
+		if (CanApplyTheme((int)ThemeProperty.TextColor))
 			TextColor = theme.TextColor;
 		_isApplyingTheme = false;
 	}
 
-	partial void OnTextColorChanging(Color? value) => TrackExplicit(ThemeProperty.TextColor, value);
+	partial void OnTextColorChanging(Color? value) => TrackExplicit((int)ThemeProperty.TextColor, value);
 }
